Replaced method addElement() by method fill() in geo operators, which should make things a lot easier. Some tests may be invalid now though. POIs are now inserted twice: as regular nodes and as POINodes.

+++ java/src/kit/route/a/lot/map/infosupply/ElementDB.java
@@ -69,7 +69,7 @@ public interface ElementDB {
      */
     public void saveToOutput(DataOutput output) throws IOException;
     
-    public POIDescription getFavoriteDescription(Coordinates pos, int detailLevel, int radius);
+    public POIDescription getFavoriteDescription(Coordinates pos, int detailLevel, float radius);
     
     public void swapNodeIDs(int id1, int id2);
